# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Subly subscription management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type SubscriptionBillingCycle = typeof SubscriptionBillingCycle[keyof typeof SubscriptionBillingCycle];


export const SubscriptionBillingCycle = {
  monthly: 'monthly',
  quarterly: 'quarterly',
  'half-yearly': 'half-yearly',
  yearly: 'yearly',
} as const;

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  cancelled: 'cancelled',
} as const;

export interface Subscription {
  id: number;
  provider: string;
  plan: string;
  category: string;
  /** @minimum 0 */
  amount: number;
  billingCycle: SubscriptionBillingCycle;
  nextRenewal: string;
  /** @minimum 0 */
  noticeDays: number;
  lastUsed: string;
  status: SubscriptionStatus;
  contractNumber: string;
  color: string;
  logoText: string;
  cancellationAddress: string;
  hotline: string;
  /** @nullable */
  priceChange?: number | null;
}

export type SubscriptionInputBillingCycle = typeof SubscriptionInputBillingCycle[keyof typeof SubscriptionInputBillingCycle];


export const SubscriptionInputBillingCycle = {
  monthly: 'monthly',
  quarterly: 'quarterly',
  'half-yearly': 'half-yearly',
  yearly: 'yearly',
} as const;

export type SubscriptionInputStatus = typeof SubscriptionInputStatus[keyof typeof SubscriptionInputStatus];


export const SubscriptionInputStatus = {
  active: 'active',
  cancelled: 'cancelled',
} as const;

export interface SubscriptionInput {
  /** @minLength 1 */
  provider: string;
  plan: string;
  category: string;
  /** @minimum 0 */
  amount: number;
  billingCycle: SubscriptionInputBillingCycle;
  nextRenewal: string;
  /** @minimum 0 */
  noticeDays: number;
  lastUsed: string;
  status: SubscriptionInputStatus;
  contractNumber: string;
  color: string;
  logoText: string;
  cancellationAddress: string;
  hotline: string;
  /** @nullable */
  priceChange?: number | null;
}

export type SubscriptionUpdateBillingCycle = typeof SubscriptionUpdateBillingCycle[keyof typeof SubscriptionUpdateBillingCycle];


export const SubscriptionUpdateBillingCycle = {
  monthly: 'monthly',
  quarterly: 'quarterly',
  'half-yearly': 'half-yearly',
  yearly: 'yearly',
} as const;

export type SubscriptionUpdateStatus = typeof SubscriptionUpdateStatus[keyof typeof SubscriptionUpdateStatus];


export const SubscriptionUpdateStatus = {
  active: 'active',
  cancelled: 'cancelled',
} as const;

export interface SubscriptionUpdate {
  /** @minLength 1 */
  provider?: string;
  plan?: string;
  category?: string;
  /** @minimum 0 */
  amount?: number;
  billingCycle?: SubscriptionUpdateBillingCycle;
  nextRenewal?: string;
  /** @minimum 0 */
  noticeDays?: number;
  lastUsed?: string;
  status?: SubscriptionUpdateStatus;
  contractNumber?: string;
  color?: string;
  logoText?: string;
  cancellationAddress?: string;
  hotline?: string;
  /** @nullable */
  priceChange?: number | null;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;