# lib/api-zod/src/generated/types/subscriptionUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Subly subscription management API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionUpdateBillingCycle } from './subscriptionUpdateBillingCycle';
import type { SubscriptionUpdateStatus } from './subscriptionUpdateStatus';

export interface SubscriptionUpdate {
  /** @minLength 1 */
  provider?: string;
  plan?: string;
  category?: string;
  /** @minimum 0 */
  amount?: number;
  billingCycle?: SubscriptionUpdateBillingCycle;
  nextRenewal?: Date;
  /** @minimum 0 */
  noticeDays?: number;
  lastUsed?: Date;
  status?: SubscriptionUpdateStatus;
  contractNumber?: string;
  color?: string;
  logoText?: string;
  cancellationAddress?: string;
  hotline?: string;
  /** @nullable */
  priceChange?: number | null;
}